requirements.txt
pycurl==7.45.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
loguru==0.7.2
tqdm==4.65.0
Jinja2==3.1.2